+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/project/renderer/extension/JavascriptProjectRendererExtension.java
@@ -1,0 +1,62 @@
+package net.splitcells.website.server.project.renderer.extension;
+
+import net.splitcells.dem.data.set.Set;
+import net.splitcells.dem.data.set.Sets;
+import net.splitcells.website.server.Config;
+import net.splitcells.website.server.project.ProjectRenderer;
+import net.splitcells.website.server.project.RenderingResult;
+
+import java.io.IOException;
+import java.nio.file.Path;
+import java.util.Optional;
+
+import static io.vertx.core.http.HttpHeaders.TEXT_HTML;
+import static java.nio.file.Files.readAllBytes;
+import static net.splitcells.dem.resource.Files.isDirectory;
+import static net.splitcells.dem.resource.Files.is_file;
+import static net.splitcells.website.server.project.RenderingResult.renderingResult;
+
+public class JavascriptProjectRendererExtension implements ProjectRendererExtension {
+    public static JavascriptProjectRendererExtension javascriptRenderer() {
+        return new JavascriptProjectRendererExtension();
+    }
+
+    private JavascriptProjectRendererExtension() {
+
+    }
+
+    @Override
+    public Optional<RenderingResult> renderFile(String path, ProjectRenderer projectRenderer, Config config) {
+        final var requestedFile = projectRenderer
+                .projectFolder()
+                .resolve("src/main/js/")
+                .resolve(path);
+        if (is_file(requestedFile)) {
+            try {
+                return Optional.of(renderingResult(readAllBytes(requestedFile), "text/javascript"));
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return Optional.empty();
+    }
+
+    @Override
+    public Set<Path> projectPaths(ProjectRenderer projectRenderer) {
+        final var projectPaths = Sets.<Path>setOfUniques();
+        final var sourceFolder = projectRenderer
+                .projectFolder()
+                .resolve("src/main/js/");
+        if (isDirectory(sourceFolder)) {
+            try {
+                java.nio.file.Files.walk(sourceFolder)
+                        .filter(java.nio.file.Files::isRegularFile)
+                        .map(file -> sourceFolder.relativize(file.getParent().resolve(file.getFileName().toString())))
+                        .forEach(projectPaths::addAll);
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+        }
+        return projectPaths;
+    }
+}